view watching and my nfts
