+++ src/views/Profile/ProfileTabs.js
@@ -1,0 +1,5 @@
+export const ProfileTabs = {
+  MyNfts: 'My NFTs',
+  Watching: 'Watching',
+  Activity: 'Activity',
+};
